requirements.txt
asgiref==3.8.1
Django==4.2.3
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
environ==1.0
gunicorn==20.1.0
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
